requirements.txt
pytest>=8.3.3
pytest-asyncio>=0.24.0
deepseek-ai>=1.0.0
